% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterparty.R
\name{counterparty}
\alias{counterparty}
\title{Create a counterparty object}
\usage{
counterparty(EAD = EAD, PD = PD, LGD = LGD, equation = equation)
}
\arguments{
\item{EAD}{}

\item{PD}{}

\item{LGD}{}

\item{equation}{}
}
\description{
Create a counterparty object
}
